Move Cog to core
