+++ core/Cog.py
