+++ core/Factory.py
@@ -1,5 +1,5 @@
 from battle.Battle import BattleCog, ToonAttack
-from Cog import Cog
+from core.Cog import Cog
 from core.Entity import BattleEntity, Entity
 from gags.Gag import Gag, get_gag_min_max_exp
 from Toon import Toon
