+++ tests/battle/test_battle_creation.py
@@ -1,9 +1,8 @@
 import pytest
 
-from core.AttackGlobals import MULTIPLIER, MULTIPLIER_DEFAULT
-
 from battle.Battle import Battle, BattleCog
-from Cog import Cog, get_random_cog
+from core.AttackGlobals import MULTIPLIER, MULTIPLIER_DEFAULT
+from core.Cog import Cog, get_random_cog
 from Toon import Toon
