+++ tests/entity/test_battle_entity_property.py
@@ -2,8 +2,8 @@ import pytest
 
 from battle.Attack import Attack
 from battle.Battle import BattleCog
-from Cog import get_random_cog
 from core.AttackGlobals import GROUP
+from core.Cog import get_random_cog
 from core.Entity import BattleEntity, Entity
 from core.Exceptions import InvalidTargetError, TargetDefeatedError
